+++ Controllers/UserController.cs
@@ -31,6 +31,7 @@ namespace WebApplication.Controllers
         [HttpPost]
         public async Task<JsonResult> GetByCredentials(CredentialsViewModel credentials)
         {
+            if(!String.IsNullOrEmpty(token))
             dynamic result = await _userService.AuthenticateUser(credentials.Email, credentials.Password);
             return Json(result);
         }
@@ -38,19 +39,23 @@ namespace WebApplication.Controllers
         [HttpPost]
         public async Task<JsonResult> Register(RegisterViewModel userData)
         {
-            dynamic result = await _userService.RegisterUser(new {
-                FirstName= userData.FirstName,
-                LastName= userData.LastName,
-                Email= userData.Email,
-                Password= userData.Password,
-                Work= userData.Work,
-                City= userData.City,
-                Country= userData.Country,
-                AcceptTerms= userData.AcceptTerms,
-                State= userData.State,
-                VoucherId= userData.VoucherCode
-            });
-            return Json(result);
+            if(!String.IsNullOrEmpty(token))
+            {
+                dynamic result = await _userService.RegisterUser(new {
+                    FirstName= userData.FirstName,
+                    LastName= userData.LastName,
+                    Email= userData.Email,
+                    Password= userData.Password,
+                    Work= userData.Work,
+                    City= userData.City,
+                    Country= userData.Country,
+                    AcceptTerms= userData.AcceptTerms,
+                    State= userData.State,
+                    VoucherId= userData.VoucherCode
+                });
+                return Json(result);
+            }
+            return Json(new { success= false, message= "Something went wrong when retrieving data, try again.", result = null });
         }        
 
         [HttpPost]
@@ -58,27 +63,24 @@ namespace WebApplication.Controllers
         {
             string token = WebApplication.Utils.Token.Get(Request.Headers);
 
-            dynamic result = await _userService.CreateUser(new User(){
-                
-               
-            //StripeToken=
-            //CustomerId=
-            //PlanId=
-            //SubscriptionId=
-            //FirstNameCard=
-            //LastNameCard=
-                 FirstName= FirstName,
-                LastName= LastName,
-                Email= Email,
-                Country= Country,
-                City= City,
-                //Site= Site,
-                Password= Password, 
-                Work= Work,
-                AcceptTerms =AcceptTerms,
-                State= 4 // Invited
-            }, token);
-            return Json(result);
+            if(!String.IsNullOrEmpty(token))
+            {
+                dynamic result = await _userService.CreateUser(new User(){
+                    FirstName= FirstName,
+                    LastName= LastName,
+                    Email= Email,
+                    Country= Country,
+                    City= City,
+                    //Site= Site,
+                    Password= Password, 
+                    Work= Work,
+                    AcceptTerms =AcceptTerms,
+                    State= 4 // Invited
+                }, token);
+                return Json(result);
+            }
+
+            return Json(new { success= false, message= "Something went wrong when retrieving data, try again.", result = null });
         }
 
         [HttpGet]
@@ -86,8 +88,12 @@ namespace WebApplication.Controllers
         {
             string token = WebApplication.Utils.Token.Get(Request.Headers);
 
-            dynamic result = await _userService.GetUserById(UserId, token);
-            return Json(result);
+            if(!String.IsNullOrEmpty(token))
+            {
+                dynamic result = await _userService.GetUserById(UserId, token);
+                return Json(result);
+            }
+            return Json(new { success= false, message= "Something went wrong when retrieving data, try again.", result = null });
         }
 
         [HttpPost]
@@ -95,16 +101,20 @@ namespace WebApplication.Controllers
         {
             string token = WebApplication.Utils.Token.Get(Request.Headers);
 
-            dynamic result = await _userService.EditUser(new User(){
-                _id= _id,
-                FirstName= FirstName,
-                LastName= LastName,
-                Email= Email,
-                City = City,
-                Country = Country
-            }, token);
+            if(!String.IsNullOrEmpty(token))
+            {
+                dynamic result = await _userService.EditUser(new User(){
+                    _id= _id,
+                    FirstName= FirstName,
+                    LastName= LastName,
+                    Email= Email,
+                    City = City,
+                    Country = Country
+                }, token);
 
-            return Json(result);
+                return Json(result);
+            }
+            return Json(new { success= false, message= "Something went wrong when retrieving data, try again.", result = null });
         }
 
         [HttpPost]
@@ -112,9 +122,12 @@ namespace WebApplication.Controllers
         {
             string token = WebApplication.Utils.Token.Get(Request.Headers);
 
-            dynamic result = await _userService.DeleteUser(_id, token);
-
-            return Json(result);
+            if(!String.IsNullOrEmpty(token))
+            {
+                dynamic result = await _userService.DeleteUser(_id, token);
+                return Json(result);
+            }
+            return Json(new { success= false, message= "Something went wrong when retrieving data, try again.", result = null });
         }
 
         [HttpGet]
